projects/bodyweight-kalman-filter/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
